+++ tools/x-env-common.sh
@@ -1,0 +1,3 @@
+export RUSTFLAGS=""
+
+export RUSTDOCFLAGS="--html-in-header ./src/_doc/header.html"