+++ tools/x-env-native.sh
@@ -1,0 +1,5 @@
+RUSTFLAGS="${RUSTFLAGS:+$RUSTFLAGS }-C target-cpu=native"
+RUSTDOCFLAGS="${RUSTDOCFLAGS:+$RUSTDOCFLAGS }-C target-cpu=native"
+
+export RUSTFLAGS
+export RUSTDOCFLAGS